More precise, platofrm-independent definition of 'Pixel32' eliminates the bug #16 (GeneralConvolution() exits with segmentation fault

+++ avxsynth/core/src/core/info.h
@@ -1548,7 +1548,7 @@ unsigned short font[][20] = {
 
 #ifndef Pixel32
 
-typedef unsigned long Pixel32;
+typedef uint32_t Pixel32;
 
 #endif
